requirements.txt
aiogram==2.9.2
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
idna==2.10
multidict==4.7.6
pyaes==1.6.1
pyasn1==0.4.8
python-dateutil==2.8.1
pytz==2020.1
rsa==4.6
six==1.15.0
Telethon==1.14.0
yarl==1.4.2
